FIX Pep8 corrections

+++ dask_cuml/tests/test_nearest_neighbors.py
@@ -17,6 +17,7 @@ from sklearn.neighbors import NearestNeighbors
 from dask.distributed import Client, wait
 from dask_cuda import LocalCUDACluster
 
+
 def test_end_to_end():
 
     cluster = LocalCUDACluster(threads_per_worker=1)
@@ -34,8 +35,10 @@ def test_end_to_end():
 
     def create_df(f, m, n):
         X = np.random.rand(m, n)
-        ret = cudf.DataFrame([(i, X[:, i].astype(np.float32)) for i in range(n)],
-                             index=cudf.dataframe.RangeIndex(f * m, f * m + m, 1))
+        ret = cudf.DataFrame([(i,
+                               X[:, i].astype(np.float32)) for i in range(n)],
+                             index=cudf.dataframe.RangeIndex(f * m,
+                                                             f * m + m, 1))
         return ret
 
     def get_meta(df):
@@ -66,19 +69,19 @@ def test_end_to_end():
     cumlNN = cumlKNN()
     cumlNN.fit(X_df)
 
-    sklNN = NearestNeighbors(metric = "sqeuclidean")
+    sklNN = NearestNeighbors(metric="sqeuclidean")
     sklNN.fit(X_pd)
 
     cuml_D, cuml_I = cumlNN.kneighbors(X_df[0:search_m-1], search_k)
     sk_D, sk_I = sklNN.kneighbors(X_pd[0:search_m], search_k)
 
-    cuml_I_nd = np.array(cuml_I.compute().as_gpu_matrix(), dtype = sk_I.dtype)
-    cuml_D_nd = np.array(cuml_D.compute().as_gpu_matrix(), dtype = sk_D.dtype)
+    cuml_I_nd = np.array(cuml_I.compute().as_gpu_matrix(), dtype=sk_I.dtype)
+    cuml_D_nd = np.array(cuml_D.compute().as_gpu_matrix(), dtype=sk_D.dtype)
 
     print(str(cuml_D_nd.dtype))
     print(str(sk_D.dtype))
 
     assert np.array_equal(cuml_I_nd, sk_I)
-    assert np.allclose(cuml_D_nd, sk_D, atol = 1e-5)
+    assert np.allclose(cuml_D_nd, sk_D, atol=1e-5)
 
     cluster.close()
